Add TableUtil.Merge function
This would add a TableUtil.Merge function, which joins two tables together using DeepCopy. Includes argument for deep copying the first table so that returned value is an entirely new table. Not 100% sure if this is the best way to implement such a feature but thought it would be a useful addition.